+++ src/background/extensionController/extensionController.js
@@ -18,8 +18,6 @@ import {
   isEquatable,
 } from "./states.js";
 
-const ONBOARDING_KEY = "mozillaVpnOnboarding";
-
 /**
  *
  * ExtensionController manages extension state and
@@ -30,9 +28,6 @@ export class ExtensionController extends Component {
   static properties = {
     state: PropertyType.Bindable,
     toggleConnectivity: PropertyType.Function,
-    nextOnboardingPage: PropertyType.Function,
-    finishOnboarding: PropertyType.Function,
-    currentOnboardingPage: PropertyType.Bindable
   };
 
   /**
@@ -54,13 +49,7 @@ export class ExtensionController extends Component {
   /** @type {VPNState} */
   clientState;
 
-  async init() {
-    this.#mCurrentOnboardingPage.value = await fromStorage(
-      browser.storage.local,
-      ONBOARDING_KEY,
-      1
-    );
-  }
+  async init() { }
 
   toggleConnectivity() {
     if (this.#mState.value.enabled) {
@@ -91,23 +80,6 @@ export class ExtensionController extends Component {
 
   get state() {
     return this.#mState.readOnly;
-  }
-
-  get currentOnboardingPage() {
-    return this.#mCurrentOnboardingPage.readOnly; // true w/ or w/o readonly
-  }
-
-  nextOnboardingPage() {
-    this.#mCurrentOnboardingPage.set(this.#mCurrentOnboardingPage.value + 1);
-  }
-
-  finishOnboarding() {
-    this.#mCurrentOnboardingPage.set(4);
-    putIntoStorage(
-      4,
-      browser.storage.local,
-      ONBOARDING_KEY
-    );
   }
 
   /**
@@ -166,6 +138,5 @@ export class ExtensionController extends Component {
   }
 
   #mState = property(new FirefoxVPNState());
-  #mCurrentOnboardingPage = property(1);
   mKeepAliveConnection = false;
 }
